move pydantic models to datamodels.py
